Switched to random port to fix port reuse bug. Not a great fix.

+++ tools/update.py
@@ -1,6 +1,9 @@
 import os
+import random
 import threading
 from time import sleep
+import sys
+
 
 def run(tsk):
     """
@@ -9,27 +12,36 @@ def run(tsk):
     """
     tsk.setup_vm()
     tsk.load(os.path.join(os.getcwd(), "remote"), "tmp")
+    sleep(3)
 
-    cmd = "start c:\\python27\\python.exe c:\\tmp\\rpcserver.py 0.0.0.0 %d vserver.EvalServer debug" % (tsk.vm.port+1)
+    tmp_port = random.randint(1025, 65000)
+
+    cmd = "start c:\\python27\\python.exe c:\\tmp\\rpcserver.py 0.0.0.0 %d vserver.EvalServer debug" % (tmp_port)
     tsk.log("TOOL:"+cmd)
     tsk.vm.launch(cmd, working_dir=r"c:\\")
     sleep(3)
 
     def close_server():
+        print 'Closing agent'
         tsk.log("Closing agent")
-        tsk.vm._guest.guest_eval("import os;os._exit(0)")
+        tsk.vm._guest.guest_eval("import os;print 'EXITING!';os._exit(0)")
 
     thread = threading.Thread(target=close_server)
     thread.start()
     thread.join(timeout=5)
+    sleep(5)
 
     tsk.log("Connecting to temporary agent")
-    tsk.vm.port += 1
-    tsk.vm.connect()
+    tsk.vm.connect(tsk.vm.addr, tmp_port)
+
+    tsk.log("Waiting for agent to be online")
+    if not tsk.vm._guest.ping():
+        tsk.log("Update failed. Could not connect to temp agent")
+        sys.exit()
 
     cmd = "rmdir /S /Q c:\\remote"
     tsk.log("TOOL:"+cmd)
-    tsk.vm.launch(cmd, working_dir=r"c:\\", verbose=True)
+    tsk.vm.launch(cmd, working_dir=r"c:\\")
     sleep(5)
 
     cmd = "xcopy /I /E c:\\tmp c:\\remote"
@@ -37,7 +49,7 @@ def run(tsk):
     tsk.vm.launch(cmd, working_dir=r"c:\\", verbose=True)
     sleep(5)
 
-    cmd = "start c:\\python27\\python.exe c:\\remote\\rpcserver.py 0.0.0.0 %d vserver.EvalServer debug" % (tsk.vm.port-1)
+    cmd = "start c:\\python27\\python.exe c:\\remote\\rpcserver.py debug"
     tsk.log("TOOL:"+cmd)
     tsk.vm.launch(cmd, working_dir=r"c:\\")
     sleep(5)
@@ -47,8 +59,7 @@ def run(tsk):
     thread.join(timeout=5)
 
     tsk.log("Connecting to new, updated agent")
-    tsk.vm.port -= 1
-    tsk.vm.connect(tsk.vm.addr, tsk.vm.port)
+    tsk.vm.connect()
 
     cmd = "rmdir /S /Q c:\\tmp"
     tsk.log("TOOL:"+cmd)
